refactor: useTimer에서 interval 계산 로직을 useTimerInterval로 분리
